+++ packages/common/src/types/index.ts
@@ -4,6 +4,7 @@
 
 export * from './Editor.types';
 export * from './ErrorHandler';
+export * from './OverridePropsParams';
 export * from './QueryEditorOptions';
 export * from './QueryNodeOptions';
 export * from './Transforms.types';