+++ packages/editor/autoformat/src/__tests__/withAutoformat/block/preFormat.spec.tsx
@@ -1,0 +1,36 @@
+/** @jsx jsx */
+
+import { jsx } from '@udecode/plate-test-utils';
+import { withReact } from 'slate-react';
+import { CONFIG } from '../../../../../../../docs/src/live/config/config';
+import { withAutoformat } from '../../../createAutoformatPlugin';
+
+jsx;
+
+const input = (
+  <editor>
+    <hul>
+      <hli>
+        <hp>
+          #
+          <cursor />
+          hello
+        </hp>
+      </hli>
+    </hul>
+  </editor>
+) as any;
+
+const output = (
+  <editor>
+    <hh1>hello</hh1>
+  </editor>
+) as any;
+
+it('should autoformat', () => {
+  const editor = withAutoformat(CONFIG.autoformat)(withReact(input));
+
+  editor.insertText(' ');
+
+  expect(input.children).toEqual(output.children);
+});