+++ packages/elements/media-embed-ui/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @udecode/plate-media-embed-ui
 
+## 6.4.1
+
+## 6.4.0
+
 ## 6.3.0
 
 ## 6.2.0