+++ packages/plate/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @udecode/plate
 
+## 6.4.2
+
+## 6.4.1
+
+## 6.4.0
+
+### Minor Changes
+
+- [#1176](https://github.com/udecode/plate/pull/1176) by [@ghingis](https://github.com/ghingis) – new packages: `line-height` and `line-height-ui`
+
 ## 6.3.0
 
 ## 6.2.0