+++ packages/ui/popover/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @udecode/plate-ui-popover
 
+## 6.4.2
+
+### Patch Changes
+
+- [#1187](https://github.com/udecode/plate/pull/1187) by [@zbeyens](https://github.com/zbeyens) – fix warning: remove `trigger` option from tippy
+
+## 6.4.1
+
+## 6.4.0
+
 ## 6.3.0
 
 ## 6.2.0